feat: Add .clang-format, .gitignore, and Makefiles
This commit introduces several new files to the project. A .clang-format
file is added to enforce a consistent coding style. A .gitignore file is
added to prevent certain files (like executables) from being tracked by Git.
Makefiles are added to the ex00, ex01, and ex02 directories to facilitate
compilation of the project. Each Makefile includes commands for building
the project, cleaning up generated files, and rebuilding the project.