feat(MutantStack): refactor class and improve iterator usage
The MutantStack class has been refactored to improve its usability and
flexibility. The class now uses a template parameter for the container
type, defaulting to std::deque. The iterator usage in the main function
has been improved to demonstrate the functionality of the class. The
class now also includes a copy constructor.